+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Uchi::Repository#title` now dynamically checks for common methods to use as the string representation of a model, ultimately falling back to `#to_s` if none are found.
 - `File` field. You can now use the File field to work with Rails ActiveStorage attachments.
 - `Image` field. Based on the `File` field it is now possible to upload images to your models and render them inline in Uchi.
+- Breadcrumb labels for links to index pages now use the title for the index page if that's been translated. You can still specify a more precise text in `repository.[name].breadcrumb.index.label`.
 
 ### Fixed